+++ packages/winnow/CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+* Support resultRecordCount and resultOffset
+
 ## [1.3.0] - 02-06-2017
 ### Added
 * Support groupBy/groupByFieldsForStatistics